pyproject.toml
[project]
name = "polymarket-hunter"
version = "0.1.0"
description = "Polymarket trading utilities and services"
readme = "README.md"
requires-python = ">=3.12"

# 👉 Keep only top-level runtime deps here (not transitive stuff like anyio/h11/idna)
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "redis>=5.0,<6.0",
  "httpx>=0.28,<0.29",
  "pydantic>=2.12,<3.0",
  "pydantic-settings>=2.11,<3.0",
  "python-dotenv>=1.0,<2.0",
  "tenacity==9.1.2",
  "web3>=7.14,<8.0",
  "websocket-client>=1.9,<2.0",
  "websockets>=15.0,<16.0",
  "pandas>=2.3,<3.0",
  "pandas-ta==0.4.71b0",
  "py-clob-client==0.28.0",
  "pytokens==0.1.10",
  "gevent>=25.9,<26.0", # keep only if you actually use it
  "greenlet>=3.2,<4.0", # keep only if you actually use it # keep only if used (httpx may be enough)
  "requests>=2.32,<3.0",
  "apscheduler>=3.11.0",
  "python-telegram-bot==22.5"
]

[dependency-groups]
dev = [
  "black>=24.0.0",
  "flake8>=7.0.0",
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pipreqs>=0.5.0",
  "pip-chill>=1.0.3"
]
